web/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
psycopg2-binary==2.9.3
pandas==1.3.5
sqlalchemy==1.4.25
openpyxl==3.0.7
numpy==1.26.4
# watchdog